requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.1.1
fastapi==0.120.0
Flask==3.1.0
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.58.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.11.3
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.0
namex==0.1.0
narwhals==1.41.0
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.6
pooch==1.8.2
protobuf==6.33.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.2.0
rpds-py==0.25.1
scikit-fuzzy==0.5.0
scikit-learn==1.6.0
scipy==1.14.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
SQLAlchemy==2.0.44
starlette==0.48.0
streamlit==1.45.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
textblob==0.19.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wordcloud==1.9.4
wrapt==2.0.0
